requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
fpdf==1.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
pdfkit==1.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
